+++ src/main/java/onlyoffice/managers/config/ConfigManagerImpl.java
@@ -25,9 +25,9 @@ import onlyoffice.managers.configuration.ConfigurationManager;
 import onlyoffice.managers.document.DocumentManager;
 import onlyoffice.managers.jwt.JwtManager;
 import onlyoffice.managers.url.UrlManager;
-import onlyoffice.model.Config;
-import onlyoffice.model.Type;
-import onlyoffice.model.editor.Mode;
+import onlyoffice.model.config.Config;
+import onlyoffice.model.config.Type;
+import onlyoffice.model.config.editor.Mode;
 import onlyoffice.utils.attachment.AttachmentUtil;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
