+++ src/main/java/onlyoffice/managers/convert/ConvertManagerImpl.java
@@ -25,7 +25,6 @@ import onlyoffice.managers.document.DocumentManager;
 import onlyoffice.managers.jwt.JwtManager;
 import onlyoffice.managers.url.UrlManager;
 import onlyoffice.model.Format;
-import onlyoffice.model.DocumentType;
 import org.apache.commons.io.IOUtils;
 import org.apache.http.HttpException;
 import org.apache.http.HttpStatus;
